+++ src/components/App.js
@@ -9,10 +9,6 @@ function App(movies) {
   const header = document.querySelector("header");
   const body = document.querySelector("body");
   const input = document.querySelector(".search_input");
-
-  var CriticallyAcclaimedTitle = "Critically Acclaimed";
-  var aZTitle = "A-Z";
-  var NewTitle = "New";
   var movieSection;
 
   const moviesContainer = document.createElement("section");
@@ -29,7 +25,7 @@ function App(movies) {
     html,
   });
 
-  // \***\*\*\*\*\*\*\***\*\*\***\*\*\*\*\*\*\***\*\*\***\*\*\*\*\*\*\***\*\*\***\*\*\*\*\*\*\***GENRES
+  // **************************************************GENRES
 
   const CriticallyAcclaimed = new MovieGenresSection(
     "critically-acclaimed",
@@ -59,23 +55,6 @@ function App(movies) {
     "ratingCount"
   );
 
-  // \***\*\*\*\*\*\*\***DONT NEED**\*\***\*\***\*\***\*\*\*\***\*\***\*\***\*\***
-  // \*\*\*or respective components
-  // AZ(aZTitle, movies, movieSection, main, header, input);
-
-  // CriticallyAcclaimed(
-  // CriticallyAcclaimedTitle,
-  // movies,
-  // movieSection,
-  // main,
-  // header,
-  // input
-  // );
-  // New(NewTitle, movies, movieSection, main, header, input);
-  // \***\*\*\*\*\*\*\***DONT NEED**\*\***\*\***\*\***\*\*\*\***\*\***\*\***\*\***
-  // \*\*\*or respective components
-
-  // \***\*\*\*\*\*\*\***\*\*\***\*\*\*\*\*\*\***\*\*\*\***\*\*\*\*\*\*\***\*\*\***\*\*\*\*\*\*\***GLOBAL
   SingleMoviePage(
     movies,
     movieSection,
@@ -85,21 +64,6 @@ function App(movies) {
     html,
     moviesContainer
   );
-
-  // Add rating colors after movies are rendered
-  const ratingTags = document.querySelectorAll(".movies_movie_text_rating");
-
-  ratingTags.forEach((tag) => {
-    if (tag.innerText >= 7) {
-      tag.classList.add("green");
-    }
-    if ((tag.innerText >= 5.4) & (tag.innerText <= 7)) {
-      tag.classList.add("orange");
-    }
-    if ((tag.innerText >= 0) & (tag.innerText <= 5.3)) {
-      tag.classList.add("red");
-    }
-  });
 }
 
 // Dont need, already rendering in the async function
