+++ src/components/ShowAllMovies.js
@@ -33,20 +33,6 @@ function ShowAllMovies(
   moviesContainer.innerHTML = allMoviesContent;
   main.appendChild(moviesContainer);
 
-  // Add rating
-  const ratingTags = document.querySelectorAll(".movies_movie_text_rating");
-  ratingTags.forEach((tag) => {
-    if (tag.innerHTML >= 7) {
-      tag.classList.add("green");
-    }
-    if ((tag.innerHTML >= 5.4) & (tag.innerHTML <= 7)) {
-      tag.classList.add("orange");
-    }
-    if ((tag.innerHTML >= 0) & (tag.innerHTML <= 5.3)) {
-      tag.classList.add("red");
-    }
-  });
-
   const newMoviesContainer = document.querySelector(".movies_container");
 
   SingleMoviePage(
